Refactor: Reorganize code into modules (#282)
