--- lib/hooks/index.ts
@@ -1,2 +1,0 @@
-export * from './useTokenAllowance';
-export * from './useDelegateAddressMap';
