+++ modules/app/components/TxInProgress.tsx
@@ -1,7 +1,7 @@
 import { Flex, Text, Box, Link } from '@theme-ui/components';
 import { Icon } from '@makerdao/dai-ui-icons';
 import TxIndicators from 'modules/app/components/TxIndicators';
-import { TXMined } from 'modules/app/types/transaction';
+import { TXMined } from 'modules/web3/types/transaction';
 import { getNetwork } from 'lib/maker';
 import { getEtherscanLink } from 'lib/utils';
