+++ modules/emergency-shutdown/components/ShutdownModal.tsx
@@ -3,9 +3,12 @@ import { useState } from 'react';
 import shallow from 'zustand/shallow';
 import getMaker, { getNetwork } from 'lib/maker';
 import { Icon } from '@makerdao/dai-ui-icons';
-import useTransactionStore, { transactionsApi, transactionsSelectors } from 'modules/app/stores/transactions';
+import useTransactionStore, {
+  transactionsApi,
+  transactionsSelectors
+} from 'modules/web3/stores/transactions';
 import { getEtherscanLink } from 'lib/utils';
-import { TXMined } from 'modules/app/types/transaction';
+import { TXMined } from 'modules/web3/types/transaction';
 import { CurrencyObject } from 'modules/app/types/currency';
 
 const ModalContent = ({
