+++ modules/polling/components/PollCreateModal.tsx
@@ -7,9 +7,12 @@ import { DialogOverlay, DialogContent } from '@reach/dialog';
 
 import { fadeIn, slideUp } from 'lib/keyframes';
 import getMaker, { getNetwork } from 'lib/maker';
-import useTransactionStore, { transactionsApi, transactionsSelectors } from 'modules/app/stores/transactions';
+import useTransactionStore, {
+  transactionsApi,
+  transactionsSelectors
+} from 'modules/web3/stores/transactions';
 import { getEtherscanLink } from 'lib/utils';
-import { TXMined } from 'modules/app/types/transaction';
+import { TXMined } from 'modules/web3/types/transaction';
 import { Poll } from 'modules/polling/types';
 
 type Props = {
