+++ modules/polling/components/PollVotingStatus.tsx
@@ -6,7 +6,7 @@ import { isActivePoll } from 'modules/polling/helpers/utils';
 import { useAllUserVotes } from 'modules/polling/hooks/useAllUserVotes';
 import useAccountsStore from 'modules/app/stores/accounts';
 import useBallotStore from 'modules/polling/stores/ballotStore';
-import useTransactionStore, { transactionsSelectors } from 'modules/app/stores/transactions';
+import useTransactionStore, { transactionsSelectors } from 'modules/web3/stores/transactions';
 import { Poll } from 'modules/polling/types';
 
 const BadgeContents = ({ hasVoted, onBallot, poll, isMined, isPending, option, ...otherProps }) => {
