+++ modules/polling/components/review/ReviewBox.tsx
@@ -8,9 +8,9 @@ import { useBreakpointIndex } from '@theme-ui/match-media';
 import { getEtherscanLink } from 'lib/utils';
 import { getNetwork } from 'lib/maker';
 import { Poll } from 'modules/polling/types';
-import { TXMined } from 'modules/app/types/transaction';
+import { TXMined } from 'modules/web3/types/transaction';
 import useBallotStore from 'modules/polling/stores/ballotStore';
-import useTransactionStore, { transactionsSelectors } from 'modules/app/stores/transactions';
+import useTransactionStore, { transactionsSelectors } from 'modules/web3/stores/transactions';
 import VotingWeight from '../VotingWeight';
 import TxIndicators from 'modules/app/components/TxIndicators';
 import PollBar from '../PollBar';
