+++ modules/web3/helpers/errors.ts
